backend/requirements.txt
Django==2.1.4
djangorestframework==3.9.1
django-cors-headers==2.4.0
